Add k8s host provisioning script
This script provisions a VM from the same image built for KVM with
packer. Some expectations of the KVM host it runs on are detailed in
comments at the top of the script. After the VM is up, the VM is
accessed via an IP provided in a virtual network used by the VM's eth0,
then the bridged IP (accessible anywhere on the network) is obtained
from eth1.

After obtaining the bridged IP, the common tools (kubectl, cf, k, helm)
are installed (SCF_BIN_DIR is set to /usr/local/bin for this)