Future-proof kube non-master node IP determination
Currently the node IP to use for DOMAIN is obtained from a jsonpath
query which lists all node IPs and takes the first one. Incidentally
this seems to be a non-master node in our deployments so far, but
nothing I've seen in the CaaSP and kubernetes documentation indicates
this order is specified in any way. Since our master nodes will always
have the 'unschedulable' property set to true, we should use a jq
filter to get the first IP from the set of nodes which *are*
shedulable.