Suppress stderr from namespace pod readiness check
When waiting for a uaa/scf deployment to be ready, the status of the
containers in the namespace is checked in a loop. If the pods don't have
the containerStatus properties yet, an error message appears in the
concourse logs which is misleading, as it doesn't actually indicate any
error.