Provide correct versions to upgrade pipeline tasks
When running upgrade tests, the old version is provided to the initial
deploy, and should be provided to pre-upgrade tests also. The old
version should *not* be provide when running tests after the upgrade.
This didn't cause issues previously because the test manifests were more
or less interchangeable, but now that the secret names (which they
reference) are changing, the tests won't be able to run if they aren't
referring to the correct secret name