Postpone setting nounset in multi-flag tasks
The cf-deploy and run-test task scripts are used by multiple tasks in
the pipeline, but will have only one 'enable' flag set for each context.
In order to avoid an error when testing to see if any of the 'enable'
flags are set, we need to move 'set -o nounset' until after this test.