Exit with status 1 when skipping teardown
This prevents the lock from being released while the kube cluster is
still in use. Presets which skip the teardown are intended for manual
investigation after a custom suite of tests, but the tester will need to
be responsible for manually tearing down CAP and releasing the lock when
they're done.