Use current branch for pipeline deploy
No longer specify the branch in the config file. The user can specify it
when deploying (with the '-b' or '--branch' flags). Otherwise, the
set-pipeline script will take the name of the current branch. If the
branch doesn't exist on the remote, or the user is in a detached head,
provide appropriate output and exit