Add teardown timeout back to pipeline definition
In commit 135c0bf511fd6a691fdf4d17546b148c3c2c90a9, we introduced a
timeout to the teardown of our cf-deploy teardown step definitions
because our teardowns would sometimes get stuck. This change was never
propagated to the other pipeline definitions due to oversight on our
part, so didn't make it into the unified pipeline definition (which was
based on the former upgrade pipeline definition).

In addition to copying this timeout back into the teardown steps, I've
added a deployments-lite preset for testing of a deploy, smoke, and
teardown.