Patch sc for default annotation only when needed
Due to https://github.com/kubernetes/kubernetes/issues/58212 (which may
only be tied to newer kubernetes versions), patching the storageclass to
set the 'default' annotation to the desired one will result in a failure
exit status when this would be a no-op. To work around this, avoid
patching an sc when it would be a no-op.