Pause jobs & unpause pipeline for new pipelines
When creating a new pipeline, all jobs should be paused automatically,
and the pipeline itself should be unpaused. Currently, the pipeline is
paused by default and if the jobs themselves aren't manually paused, all
jobs can automatically start when the pipeline is unpaused.

By pausing all jobs by default and unpausing the pipeline, jobs for
newly created pipelines won't start automatically, and a user can
manually unpause a job when they want to run it.

If the pipeline which is being set already existed, update the pipeline
according to the new parameters, but don't pause or unpause the pipeline
or its jobs.