Temporary workaround for brains psp issue
Right now we have a 'test-brain' service-account which the manifest for
our brains test suite is associated with. Unfortunately this
service-account wasn't referenced in the clusterrolebinding for our psp
workaround, so the pod doesn't have the right privileges for its NFS
test. This adds the 'test-brain' service-account to the psp workaround
clusterrolebinding, and when running the brains test suite in the
pipeline, will kubectl apply the workaround again if the 'test-brain'
serviceaccount isn't already included in the existing clusterrolebinding