Only print recurring failures when CATs has run 5x
If CATs passes on the first run, recurring_failures never gets set, and
the check for recurring_failures when printing results from CATs yields
an error due to the errexit shell option. Since recurring_failures is
empty when passing, but can also be "unset" (when re-running CATs, to
signal current_failures to be copied to recurring_failures without the
check for overlap), it has a few possible states by the time we get to
the final check. To avoid checking it when one of those states can
actually be an unset variable, we can check the value of CATS_RERUN
first, because it will always equal 5 when there are recurring failures.