Generate single-use ssh key for AKS node access
This commit introduces the generation of a single-use ssh keypair for
AKS node access. Previously there has been no easy, secure way to ssh to
the nodes, since they're only accessible via their internal IPs. Based on
instructions from https://docs.microsoft.com/en-us/azure/aks/ssh, the
recommended way to gain AKS node access is via ssh to its internal IP
from a pod, but doing this requires copying the private key to the pod
with 'kubectl cp'. This means only the user who deployed the cluster
*should* have access to the private key, and furthermore, by copying the
key to a pod, they're exposing it to anyone with cluster access via the
kube config, which is shared with the CAP QA team, and in the event this
config file falls into the hands of an attacker, we want to avoid
putting anything like private, reusable ssh keys which grant greater
organizational access, into the cluster pods.