Push klog to s3 after github & email notifications
In some situations where the deploy fails, a klog will not be generated.
In these cases, klog/target_url will link to the concourse build URL
instead of a klog, but the s3 push will still fail because nothing will
match the resource's file param pattern. In order to ensure
notifications are still sent/set, they we should push the klog to s3
after.

Also, the email generation task should only construct the klog section
when a klog file exists, not when the target_url file exists