Add qa-tools/skuba-deploy.sh and docker/skuba
skuba-deploy.sh allows to bootstrap clusters with skuba, configure the nodes in
the cluster, run commands on them, etc.

It uses a docker image containing the correct version of skuba, terraform, and
ssh from the CaasP repos. The image can be configured to install CaasP from
devel, updates, staging or product.

The image is punctured: it reads the ssh agent envs from the host, uid and gid,
and creates files in the host when running skuba and terraform. This allows for
running the caasp binaries locally without doing it from a sle15sp1 image.